+++ api/v1/views/cities.py
@@ -1,77 +1,71 @@
 #!/usr/bin/python3
-"""Module cities"""
-from models.city import City
-from models.state import State
-from models import storage
+"""objects that handles all default RestFul API actions for Amenities"""
+
+from flask import abort, request, jsonify, make_response
 from api.v1.views import app_views
-from flask import *
+from models import storage
+from models import amenity
+from models.amenity import Amenity
 
 
-@app_views.route('/states/<state_id>/cities', methods=['GET'],
+@app_views.route("/amenities", strict_slashes=False)
+def get_amenity():
+    """Method for amenity"""
+    new_list = []
+    for amenity in storage.all(Amenity).values():
+        new_list.append(amenity.to_dict())
+    return jsonify(new_list)
+
+
+@app_views.route("/amenities/<string:amenity_id>", strict_slashes=False)
+def one_amenity(amenity_id):
+    """Method for one amenity"""
+    amenity = storage.get(Amenity, amenity_id)
+    if amenity is None:
+        abort(404)
+    return jsonify(amenity.to_dict())
+
+
+@app_views.route("/amenities/<string:amenity_id>", methods=["DELETE"],
                  strict_slashes=False)
-def get_cities(state_id):
-    """Retrieves the list of all City objects of a State"""
-    list_cities = []
-    state = storage.get(State, state_id)
-    if not state:
-        return make_response(jsonify({"error": "Not found"}), 404)
-    for city in state.cities:
-        list_cities.append(city.to_dict())
-    return jsonify(list_cities)
-
-
-@app_views.route('/cities/<city_id>/', methods=['GET'], strict_slashes=False)
-def get_city(city_id):
-    """Retrieves a City object"""
-    city = storage.get(City, city_id)
-    if not city:
-        return make_response(jsonify({"error": "Not found"}), 404)
-    return jsonify(city.to_dict())
-
-
-@app_views.route("/cities/<city_id>", strict_slashes=False, methods=["DELETE"])
-def delete_city(city_id=None):
-    """Deletes a City object"""
-    city = storage.get(City, city_id)
-    if not city:
-        return make_response(jsonify({"error": "Not found"}), 404)
-    storage.delete(city)
+def amenity_delete(amenity_id):
+    """Method that deletes a amenity object"""
+    amenity = storage.get(Amenity, amenity_id)
+    if not amenity:
+        abort(404)
+    storage.delete(amenity)
     storage.save()
-    return make_response(jsonify({}), 200)
+    return make_response(jsonify(({})), 200)
 
 
-@app_views.route("/states/<state_id>/cities", strict_slashes=False,
-                 methods=["POST"])
-def post_city(state_id):
-    """Creates a City"""
-    state = storage.get(State, state_id)
-    data = request.get_json(force=True, silent=True)
-    if not state:
-        return make_response(jsonify({"error": "Not found"}), 404)
+@app_views.route('/amenities', methods=['POST'], strict_slashes=False)
+def amenity_post():
+    """Method that creates a amenity"""
+    data = request.get_json()
     if not data:
-        return make_response(jsonify({"error": "Not a JSON"}), 400)
+        abort(400, description="Not a JSON")
     if "name" not in data:
-        return make_response(jsonify({"error": "Missing name"}), 400)
-    s = City(**data)
-    s.state_id = state.id
-    s.save()
-    return make_response(jsonify(s.to_dict()), 201)
+        abort(400, description="Missing name")
+    instance = Amenity(**data)
+    instance.save()
+    return make_response(jsonify(instance.to_dict()), 201)
 
 
-@app_views.route("/cities/<city_id>", strict_slashes=False, methods=["PUT"])
-def put_city(city_id=None):
-    """Updates a City object"""
-    city = storage.get(City, city_id)
-    data = request.get_json(force=True, silent=True)
+@app_views.route("/amenities/<string:amenity_id>", methods=['PUT'],
+                 strict_slashes=False)
+def amenity_put(amenity_id):
+    """Method that puts a amenity"""
+    amenity = storage.get(Amenity, amenity_id)
+    data = request.get_json()
+    if not amenity:
+        abort(404)
     if not data:
-        return make_response(jsonify({"error": "Not a JSON"}), 400)
+        abort(400, description="Not a JSON")
 
-    ignore_keys = ["id", "state_id", "created_at", "updated_at"]
-    if not city:
-        return make_response(jsonify({"error": "Not found"}), 404)
-    else:
-        for key, value in data.items():
-            if key not in ignore_keys:
-                setattr(city, key, value)
+    ignore = ['id', 'created_at', 'updated_at']
+
+    for key, value in data.items():
+        if key not in ignore:
+            setattr(amenity, key, value)
         storage.save()
-        return make_response(jsonify(city.to_dict()), 200)
+        return make_response(jsonify(amenity.to_dict()), 200)
